Separate routes into different files, control by config
