+++ distributed/http/statics.py
@@ -1,0 +1,10 @@
+from tornado import web
+import os
+
+routes = [
+    (
+        r"/statics/(.*)",
+        web.StaticFileHandler,
+        {"path": os.path.join(os.path.dirname(__file__), "static")},
+    ),
+]
